+++ modules/vpc/igw.tf
@@ -5,6 +5,7 @@
 
 # https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/internet_gateway
 resource "aws_internet_gateway" "igw" {
+  count  = var.enable_igw ? 1 : 0
   vpc_id = aws_vpc.main.id
   tags = merge(var.tags, {
     Name = "${var.name}-igw"
@@ -13,10 +14,11 @@ resource "aws_internet_gateway" "igw" {
 
 # https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/route_table
 resource "aws_route_table" "public" {
+  count  = var.enable_igw ? 1 : 0
   vpc_id = aws_vpc.main.id
   route {
     cidr_block = "0.0.0.0/0"
-    gateway_id = aws_internet_gateway.igw.id
+    gateway_id = aws_internet_gateway.igw[0].id
   }
   route {
     cidr_block = var.cidr
@@ -29,7 +31,7 @@ resource "aws_route_table" "public" {
 
 # https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/main_route_table_association
 resource "aws_route_table_association" "public" {
-  for_each       = { for i, subnet in aws_subnet.public : i => subnet }
+  for_each       = var.enable_igw ? { for i, subnet in aws_subnet.public : i => subnet } : {}
   subnet_id      = each.value.id
-  route_table_id = aws_route_table.public.id
+  route_table_id = aws_route_table.public[0].id
 }