+++ modules/vpc/nat.tf
@@ -5,6 +5,7 @@
 
 # https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/eip
 resource "aws_eip" "nat" {
+  count  = var.enable_nat ? 1 : 0
   domain = "vpc"
   tags = merge(var.tags, {
     Name = "${var.name}-nat-ip"
@@ -14,7 +15,8 @@ resource "aws_eip" "nat" {
 
 # https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/nat_gateway
 resource "aws_nat_gateway" "ngw" {
-  allocation_id     = aws_eip.nat.id
+  count             = var.enable_nat ? 1 : 0
+  allocation_id     = aws_eip.nat[0].id
   connectivity_type = "public"
   subnet_id         = aws_subnet.public[0].id
   tags = merge(var.tags, {
@@ -26,10 +28,11 @@ resource "aws_nat_gateway" "ngw" {
 # Instead of creating a Private-NAT Route Table we re-use the default
 # https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/default_route_table
 resource "aws_default_route_table" "private" {
+  count                  = var.enable_nat ? 1 : 0
   default_route_table_id = aws_vpc.main.default_route_table_id
   route {
     cidr_block     = "0.0.0.0/0"
-    nat_gateway_id = aws_nat_gateway.ngw.id
+    nat_gateway_id = aws_nat_gateway.ngw[0].id
   }
   route {
     cidr_block = var.cidr
@@ -42,7 +45,7 @@ resource "aws_default_route_table" "private" {
 
 # https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/main_route_table_association
 resource "aws_route_table_association" "private" {
-  for_each       = { for i, subnet in aws_subnet.private : i => subnet }
+  for_each       = var.enable_nat ? { for i, subnet in aws_subnet.private : i => subnet } : {}
   subnet_id      = each.value.id
-  route_table_id = aws_default_route_table.private.id
+  route_table_id = aws_default_route_table.private[0].id
 }